feat: add comprehensive tests for shell command permission checks, including allowlisting, blocklisting, and injection prevention.
